requirements.txt
pillow
pillow-heif
tensorflow
opencv-python
scipy
pandas